+++ src/app/(player)/Sidebar.tsx
