+++ src/app/(player)/playlist/ActionsMenu.tsx
@@ -28,7 +28,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "@/components/ui/button";
-import { type Data } from "@/components/TrackTable/playlistColumns";
+import { type Data } from "@/app/(player)/playlist/PlaylistTable";
 
 type ActionsMenuProps = {
   album: Data,
