+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import { type Metadata } from "next";
 import { Inter } from "next/font/google";
 import "@/styles/globals.css";
-import { ThemeProvider } from "@/app/theme-provider";
+import { ThemeProvider } from "@/app/ThemeProvider";
 
 const inter = Inter({
   subsets: ["latin"],
