+++ src/components/PlaybackBar/CurrentlyPlaying.tsx
